python_object_detection_gui/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=9.5.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0